WebScraper_V1.1/requirements.txt
selenium==4.0.0
selenium-wire==4.5.4
requests
psutil
packaging
blinker<1.8.0
pyOpenSSL==20.0.1
cryptography==3.3.2
fake_useragent
psycopg2-binary